+++ fit.py
@@ -20,7 +20,7 @@ from scipy.optimize import fmin_bfgs
 from joblib import Parallel, delayed
 from scipy import interpolate
 import matplotlib.pyplot as plt
-sys.path.append("D:\Git\ExpSIMCE")
+sys.path.append("C:/Users\Patricio De Araya\Dropbox\LocalRA\Local_teacherGITnewmodel")
 #sys.path.append("D:\Git\WageError")
 #import gridemax
 import time
@@ -37,15 +37,15 @@ from openpyxl import load_workbook
 np.random.seed(123)
 
 #betas_nelder  = np.load("D:\Git\ExpSIMCE/betasopt_model_RA3.npy")
-betas_nelder  = np.load("/Users/jorge-home/Dropbox/Research/teachers-reform/codes/teachers/betasopt_model_v23.npy")
+betas_nelder  = np.load("C:/Users\Patricio De Araya\Dropbox\LocalRA\Local_teacherGITnewmodel/betasopt_model_v2023.npy")
 
 #moments_vector = np.load("D:\Git\ExpSIMCE/moments.npy")
-moments_vector = np.load("/Users/jorge-home/Dropbox/Research/teachers-reform/codes/teachers/moments.npy")
+moments_vector = np.load("C:/Users\Patricio De Araya\Dropbox\LocalRA\Local_teacherGITnewmodel/moments_v2023.npy")
 
 #ajhdsajk = moments_vector[0,1]
 
 #data = pd.read_stata('D:\Git\ExpSIMCE/data_pythonpast.dta')
-data = pd.read_stata('/Users/jorge-home/Dropbox/Research/teachers-reform/codes/teachers/data_pythonpast.dta')
+data = pd.read_stata('C:/Users\Patricio De Araya\Dropbox\LocalRA\Local_teacherGITnewmodel/data_pythonpast_v2023.dta')
 
 
 
@@ -86,6 +86,8 @@ typeSchool = np.array(data['typeschool'])
 
 # Priority #
 priotity = np.array(data['por_priority'])
+
+priority_aep = np.array(data['priority_aep'])
 
 rural_rbd = np.array(data['rural_rbd'])
 
@@ -157,13 +159,13 @@ priori = [pri[0]/dolar, pri[1]/dolar]
 param0 = parameters.Parameters(alphas,betas,gammas,hw,porc,pro,pol,AEP,priori)
 
 model = util.Utility(param0,N,p1_0,p2_0,years,treatment,typeSchool,HOURS,p1,p2,catPort,catPrueba,
-                     TrameI,priotity,rural_rbd,locality)
+                     TrameI,priotity,rural_rbd,locality, priority_aep)
 
 modelSD = sd.SimData(N,model)
 
 
 #ses_opt = np.load("D:\Git\ExpSIMCE/ses_model.npy")
-ses_opt = np.load("/Users/jorge-home/Dropbox/Research/teachers-reform/codes/teachers/ses_model.npy")
+ses_opt = np.load("C:/Users\Patricio De Araya\Dropbox\LocalRA\Local_teacherGITnewmodel/ses_model_v2023.npy")
 w_matrix = np.zeros((ses_opt.shape[0],ses_opt.shape[0]))
 
 #var_cov = np.load("/Users/jorge-home/Dropbox/Research/teachers-reform/codes/teachers/var_cov.npy")
@@ -171,11 +173,12 @@ w_matrix = np.zeros((ses_opt.shape[0],ses_opt.shape[0]))
 
 for j in range(ses_opt.shape[0]):
     w_matrix[j,j] = ses_opt[j]**(-2)
+    
 
 output_ins = est.estimate(N, years,param0, p1_0,p2_0,treatment, \
-                 typeSchool,HOURS,p1,p2,catPort,catPrueba,TrameI,priotity,rural_rbd,locality, 
+                 typeSchool,HOURS,p1,p2,catPort,catPrueba,TrameI,priotity,rural_rbd,locality, priority_aep, \
                  w_matrix,moments_vector)
-
+    
 
 corr_data = output_ins.simulation(50,modelSD)
 print(corr_data)
@@ -202,11 +205,12 @@ beta0 = np.array([param0.alphas[0][0],
 print(beta0)
 
 qw = output_ins.objfunction(beta0)
+print(qw) 
 
 ##### PYTHON TO EXCEL #####
 
 #wb = load_workbook('D:\Git\ExpSIMCE/Outcomes.xlsx')
-wb = load_workbook('/Users/jorge-home/Dropbox/Research/teachers-reform/codes/teachers/Outcomes.xlsx')
+wb = load_workbook('C:/Users\Patricio De Araya\Dropbox\LocalRA\Local_teacherGITnewmodel/Outcomes_v2023.xlsx')
 sheet = wb["data"]
 
 sheet['C5'] = 'Mean Portfolio'
@@ -279,7 +283,7 @@ sheet['D22'] = corr_data['Estimation SIMCE vs Experience']
 
 
 #wb.save('D:\Git\ExpSIMCE/Outcomes.xlsx')
-wb.save('/Users/jorge-home/Dropbox/Research/teachers-reform/codes/teachers/Outcomes.xlsx')
+wb.save('C:/Users\Patricio De Araya\Dropbox\LocalRA\Local_teacherGITnewmodel/Outcomes_v2023.xlsx')
